requirements.txt
pika
psycopg2-binary
python-dotenv
pytz
tzdata
requests
fastapi
uvicorn